# backend/guia_cerrado/docs/swagger.json
{
...
  "info": {
    "description": "Documentação da API do Guia Cerrado",
    "termsOfService": "/tos",
    "title": "Guia Cerrado API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/recommendations": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "culture": {
                  "description": "Nome da cultura agrícola",
                  "example": "Soja",
                  "type": "string"
                },
                "lat": {
                  "description": "Latitude da localização (deve estar dentro do Cerrado)",
                  "example": -15.7801,
                  "format": "float",
                  "type": "number"
                },
                "lng": {
                  "description": "Longitude da localização (deve estar dentro do Cerrado)",
                  "example": -47.9292,
                  "format": "float",
                  "type": "number"
                },
                "problem_description": {
                  "description": "Descrição do problema observado",
                  "example": "Folhas com manchas amarelas e queda prematura",
                  "type": "string"
                },
                "production_stage": {
                  "description": "Estágio de produção da cultura",
                  "example": "Floração",
                  "type": "string"
                }
              },
              "required": [
                "lat",
                "lng",
                "culture",
                "production_stage",
                "problem_description"
              ],
              "type": "object"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Relatório técnico gerado com sucesso",
            "schema": {
              "properties": {
                "cerrado_data": {
                  "description": "Dados ambientais da região do Cerrado para as coordenadas fornecidas",
                  "type": "object"
                },
                "culture": {
                  "example": "Soja",
                  "type": "string"
                },
                "problem_reported": {
                  "example": "Folhas com manchas amarelas e queda prematura",
                  "type": "string"
                },
                "production_stage": {
                  "example": "Floração",
                  "type": "string"
                },
                "recommendations": {
                  "items": {
                    "properties": {
                      "description": {
                        "description": "Descrição em HTML gerada a partir de Markdown",
                        "example": "<p>A coloração amarelada nas folhas indica...</p>",
                        "type": "string"
                      },
                      "title": {
                        "example": "Possível deficiência de nitrogênio",
                        "type": "string"
                      },
                      "type": {
                        "enum": [
                          "diagnóstico",
                          "soluções",
                          "medidas preventivas"
                        ],
                        "example": "diagnóstico",
                        "type": "string"
                      }
                    },
                    "type": "object"
                  },
                  "type": "array"
                }
              },
              "type": "object"
            }
          },
          "400": {
            "description": "Requisição inválida",
            "schema": {
              "examples": {
                "campo_ausente": {
                  "value": {
                    "error": "Campo 'lat' ausente"
                  }
                },
                "coordenadas_ausentes": {
                  "value": {
                    "error": "Latitude ou longitude ausentes"
                  }
                },
                "fora_do_cerrado": {
                  "value": {
                    "error": "Localização fora da região do Cerrado"
                  }
                }
              },
              "properties": {
                "error": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        },
        "summary": "Gera recomendações técnicas para culturas no Cerrado",
        "tags": [
          "Recommendation"
        ]
      }
    },
    "/recommendations_fake": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "description": "Endpoint de teste que valida os dados de entrada, verifica se a localização pertence ao Cerrado e retorna uma resposta simulada com recomendações técnicas. Útil para desenvolvimento de frontend, testes de integração e validação do contrato da API.\n",
        "parameters": [
          {
            "description": "Dados necessários para gerar a recomendação simulada",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "culture": {
                  "description": "Cultura agrícola informada pelo usuário",
                  "example": "eucalipto",
                  "type": "string"
                },
                "lat": {
                  "description": "Latitude da área analisada",
                  "example": -16.6869,
                  "format": "float",
                  "type": "number"
                },
                "lng": {
                  "description": "Longitude da área analisada",
                  "example": -49.2648,
                  "format": "float",
                  "type": "number"
                },
                "problem_description": {
                  "description": "Problema observado na cultura",
                  "example": "caule muito fino",
                  "type": "string"
                },
                "production_stage": {
                  "description": "Estágio atual da produção",
                  "example": "plantio",
                  "type": "string"
                }
              },
              "required": [
                "lat",
                "lng",
                "culture",
                "production_stage",
                "problem_description"
              ],
              "type": "object"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Relatório técnico simulado retornado com sucesso",
            "examples": {
              "application/json": {
                "biome_location": "Cerrado",
                "cerrado_data": {},
                "culture": "eucalipto",
                "problem_reported": "caule muito fino",
                "production_stage": "plantio",
                "recommendations": [
                  {
                    "description": "O caule fino do eucalipto pode ser resultado de várias condições ambientais e de manejo.",
                    "title": "Análise das Possíveis Causas",
                    "type": "diagnóstico"
                  },
                  {
                    "description": "Recomenda-se adubação balanceada, irrigação suplementar e controle da vegetação competidora.",
                    "title": "Soluções Práticas para o Problema",
                    "type": "soluções"
                  },
                  {
                    "description": "Realizar análise periódica do solo e adotar práticas de conservação pode reduzir a recorrência do problema.",
                    "title": "Prevenção de Recorrência",
                    "type": "medidas preventivas"
                  }
                ]
              }
            },
            "schema": {
              "properties": {
                "biome_location": {
                  "description": "Bioma associado à localização analisada",
                  "example": "Cerrado",
                  "type": "string"
                },
                "cerrado_data": {
                  "description": "Dados ambientais e territoriais calculados para as coordenadas fornecidas",
                  "type": "object"
                },
                "culture": {
                  "example": "eucalipto",
                  "type": "string"
                },
                "problem_reported": {
                  "example": "caule muito fino",
                  "type": "string"
                },
                "production_stage": {
                  "example": "plantio",
                  "type": "string"
                },
                "recommendations": {
                  "items": {
                    "properties": {
                      "description": {
                        "example": "O caule fino do eucalipto pode ser resultado de várias condições ambientais e de manejo.",
                        "type": "string"
                      },
                      "title": {
                        "example": "Análise das Possíveis Causas",
                        "type": "string"
                      },
                      "type": {
                        "enum": [
                          "diagnóstico",
                          "soluções",
                          "medidas preventivas"
                        ],
                        "example": "diagnóstico",
                        "type": "string"
                      }
                    },
                    "required": [
                      "type",
                      "title",
                      "description"
                    ],
                    "type": "object"
                  },
                  "type": "array"
                }
              },
              "type": "object"
            }
          },
          "400": {
            "description": "Erro de validação da requisição",
            "examples": {
              "application/json": {
                "error": "Campo 'lat' ausente"
              }
            },
            "schema": {
              "properties": {
                "error": {
                  "example": "Campo 'lat' ausente",
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        },
        "summary": "Gera uma recomendação técnica simulada para testes",
        "tags": [
          "Recommendation"
        ]
      }
    },
...
  },
  "swagger": "2.0"
}